bank_chatbot/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary>=2.9.0
alembic==1.12.1

# Redis
redis[hiredis]>=5.0.0

# OpenAI
openai>=1.3.0

# HTTP Client
httpx>=0.25.2

# Environment
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.6

